Localize TUI app tests
